refactor(web): modularize InferenceLatency page components
- Extracted new components (BreachBanner, SourceAdvisory, FtSearchTrendPanel, BucketTile) from InferenceLatency.tsx for better encapsulation and maintainability.
- Simplified InferenceLatency page by composing it with the modularized components.
- Improved trend indexing band handling in FtSearchTrendPanel with end time consistency.
- Updated bucket sorting logic for better readability and performance.
